requirements/base.txt
Django==3.0.11
djangorestframework==3.12.1
django-cors-headers==3.5.0
django-extensions==3.0.9
mysqlclient==2.0.1
faker-commerce==1.0.4
pytest==8.3.3
pytest-cov==5.0.0
pytest-django==4.9.0
pytest-forked==1.6.0
pytest-xdist==3.6.1
factory-boy==3.2.1